Use ConfigOpts.find_file() for policy and paste
Use cfg's new helper method to find the PasteDeploy and policy config
files. The basic behavior is "look alongside the config file" with a
fall back to the standard default config paths.

In the fairly obscure case where no --config-file is found we use an
explicit default paste file taking to account the fact that all the
glance apps have their paste config in a single file.

Add some new debug logging to make it easier to see which config files
are being loaded.

Fixes bug 992319

Change-Id: I978c4a8cb6ec6d12aaf746af511ccb572c34b7ee