wsgi: do not respawn on missing eventlet hub
Eventlet 'poll' hub is not available on OSX, wsgi.Server workers have no
chance to start but the error was not caught and they were respawned
endlessly. This patch makes these workers not to be respawned.

This has the nice side-effect of avoiding the test suite to get stuck in
test_multiprocessing on OSX, filling syslog.

Change-Id: I565d21f5c6861102e96d6953ace29cde57e8ec38